pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tyler-agent"
version = "0.4.0"
description = "Tyler: A framework for manifesting AI agents with a complete lack of conventional limitations"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "CC BY-NC 4.0"}
authors = [
    {name = "adamwdraper"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "License :: Free for non-commercial use",
]
dependencies = [
    "litellm>=1.60.2",
    "openai>=1.61.0",
    "tiktoken>=0.8.0",
    "pdf2image>=1.17.0",
    "pypdf>=5.3.0",
    "python-magic>=0.4.0",
    "pillow>=11.0.0",
    "SQLAlchemy>=2.0.36",
    "greenlet>=3.1.1",
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.21.0",
    "psycopg2-binary>=2.9.9",
    "aiohttp>=3.11.11",
    "httpx>=0.27.2",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.1",
    "click>=8.1.8",
    "pydantic>=2.10.4",
    "backoff>=2.2.1",
    "uuid_utils>=0.10.0",
    "weave>=0.51.32",
    "wandb>=0.19.1",
    "slack_sdk>=3.34.0",
    "huggingface-hub>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "coverage>=7.6.10",
    "pip-tools>=7.4.1",
    "pipdeptree>=2.25.0",
]

[project.urls]
Homepage = "https://github.com/adamwdraper/tyler"
Documentation = "https://github.com/adamwdraper/tyler#readme"
Repository = "https://github.com/adamwdraper/tyler"
"Bug Tracker" = "https://github.com/adamwdraper/tyler/issues"

[project.scripts]
tyler-db = "tyler.database.cli:main"
tyler-chat = "tyler.cli.chat:main"

[tool.hatch.build.targets.wheel]
packages = ["tyler"]

[tool.hatch.build]
include = [
    "tyler/**/*.py",
    "tyler/database/migrations/alembic.ini",
    "tyler/database/migrations/script.py.mako",
    "tyler/database/migrations/env.py",
    "tyler/database/migrations/versions/*.py",
]
exclude = [
    "**/.env",
    "**/*.pyc",
    "**/__pycache__",
    "**/*.db",
    "**/*.sqlite",
    "**/data",
    "**/.coverage",
    "**/htmlcov",
    "**/dist",
    "**/build",
    "**/*.egg-info",
] 